+++ spec/ruby/core/objectspace/weakmap/each_pair_spec.rb
@@ -3,8 +3,9 @@ require_relative 'shared/members'
 require_relative 'shared/each'
 
 describe "ObjectSpace::WeakMap#each_pair" do
-  it_behaves_like :members, ->(map) { a = []; map.each_pair{ |k,v| a << "#{k}#{v}" }; a }, %w[Ax By]
+  it_behaves_like :weakmap_members, ->(map) { a = []; map.each_pair{ |k,v| a << "#{k}#{v}" }; a }, %w[Ax By]
+end
 
-  before(:all) { @method_name = :each_pair }
-  it_should_behave_like :each
+describe "ObjectSpace::WeakMap#each_key" do
+  it_behaves_like :weakmap_each, :each_pair
 end