+++ spec/ruby/core/objectspace/weakmap/each_value_spec.rb
@@ -3,8 +3,9 @@ require_relative 'shared/members'
 require_relative 'shared/each'
 
 describe "ObjectSpace::WeakMap#each_value" do
-  it_behaves_like :members, ->(map) { a = []; map.each_value{ |k| a << k }; a }, %w[x y]
+  it_behaves_like :weakmap_members, ->(map) { a = []; map.each_value{ |k| a << k }; a }, %w[x y]
+end
 
-  before(:all) { @method_name = :each_value }
-  it_should_behave_like :each
+describe "ObjectSpace::WeakMap#each_key" do
+  it_behaves_like :weakmap_each, :each_value
 end