+++ spec/ruby/core/objectspace/weakmap/include_spec.rb
@@ -2,5 +2,5 @@ require_relative '../../../spec_helper'
 require_relative 'shared/include'
 
 describe "ObjectSpace::WeakMap#include?" do
-  it_behaves_like :include?, :include?
+  it_behaves_like :weakmap_include?, :include?
 end