+++ spec/ruby/core/objectspace/weakmap/key_spec.rb
@@ -2,5 +2,5 @@ require_relative '../../../spec_helper'
 require_relative 'shared/include'
 
 describe "ObjectSpace::WeakMap#key?" do
-  it_behaves_like :include?, :key?
+  it_behaves_like :weakmap_include?, :key?
 end