+++ spec/ruby/core/objectspace/weakmap/length_spec.rb
@@ -2,5 +2,5 @@ require_relative '../../../spec_helper'
 require_relative 'shared/size'
 
 describe "ObjectSpace::WeakMap#length" do
-  it_behaves_like :size, :length
+  it_behaves_like :weakmap_size, :length
 end