+++ spec/ruby/core/objectspace/weakmap/member_spec.rb
@@ -2,5 +2,5 @@ require_relative '../../../spec_helper'
 require_relative 'shared/include'
 
 describe "ObjectSpace::WeakMap#member?" do
-  it_behaves_like :include?, :member?
+  it_behaves_like :weakmap_include?, :member?
 end