+++ spec/ruby/core/objectspace/weakmap/shared/size.rb
@@ -1,4 +1,4 @@
-describe :size, shared: true do
+describe :weakmap_size, shared: true do
   it "is correct" do
     map = ObjectSpace::WeakMap.new
     key1, key2 = %w[a b].map(&:upcase)