+++ spec/ruby/core/objectspace/weakmap/size_spec.rb
@@ -2,5 +2,5 @@ require_relative '../../../spec_helper'
 require_relative 'shared/size'
 
 describe "ObjectSpace::WeakMap#size" do
-  it_behaves_like :size, :size
+  it_behaves_like :weakmap_size, :size
 end