+++ src/main/java/org/truffleruby/core/GCNodes.java
@@ -42,7 +42,7 @@ public abstract class GCNodes {
      * much more likely to actually trigger GC than calling {@link System#gc()} or Ruby's {@link GCStartPrimitiveNode}.
      *
      * <p>
-     * In particular, this attempts to trigger the GC. by waiting until a weak reference has been cleared.
+     * In particular, this attempts to trigger the GC by waiting until a weak reference has been cleared.
      *
      * <p>
      * Note that even when GC is triggered, there is not guarantee that the all the garbage has been cleared or all the
@@ -54,7 +54,8 @@ public abstract class GCNodes {
         @Specialization
         protected Object force() {
             getContext().getMarkingService().queueMarking();
-            Object key = new Object(), value = new Object();
+            Object key = new Object();
+            Object value = new Object();
 
             // NOTE(norswap, 16 Apr 20): We could have used a WeakReference here, but the hope is that the extra
             // indirection will prevent the compiler to optimize this method away (assuming JIT compilation, and